Rename react-auth0-spa.js to Auth0Provider.js
